Cleaned up to use react context
